Update Audit Logs

+++ apps/builder/src/features/logs/delete/delete-logs-queries.ts
@@ -4,9 +4,9 @@ import { authActionClient } from "@/lib/safe-action";
 import { findChatbotOrFail } from "@/lib/user-permissions";
 import { prisma } from "@ahachat.ai/database";
 import { User } from "@prisma/client";
+import { returnValidationErrors } from "next-safe-action";
 import { revalidateTag } from "next/cache";
-import { z } from "zod";
-import { deleteLogSchema } from "./delete-logs-schema";
+import { DeleteLogSchema, deleteLogSchema } from "./delete-logs-schema";
 
 export const deleteLogAction = authActionClient
   .schema(deleteLogSchema)
@@ -15,34 +15,29 @@ export const deleteLogAction = authActionClient
     parsedInput,
   }: {
     ctx: { user: User };
-    parsedInput: z.infer<typeof deleteLogSchema>;
+    parsedInput: DeleteLogSchema;
   }) => {
-    try {
 
-      console.log("Xem có id log không? ", parsedInput.ids)
+    const { chatbot } = await findChatbotOrFail(ctx.user.id, parsedInput.chatbotId);
 
-      const { chatbot } = await findChatbotOrFail(ctx.user.id, parsedInput.chatbotId);
-
-      await prisma.log.deleteMany({
-        where: {
-          id: {
-            in: parsedInput.ids,
-          },
-          chatbotId: chatbot.id,
-        },
-      });
-
-      revalidateTag("logs");
-
-      return {
-        successful: true,
-      };
-    } catch (err) {
-      console.error(err);
-
-      return {
-        successful: false,
-        error: err,
-      };
+    if (!chatbot) {
+      return returnValidationErrors(deleteLogSchema, {
+        _errors: ["Delete Exception"],
+      })
     }
+
+    await prisma.log.deleteMany({
+      where: {
+        id: {
+          in: parsedInput.ids,
+        },
+        chatbotId: chatbot.id,
+      },
+    });
+
+    revalidateTag(`${ctx.user.id}#logs`);
+
+    return {
+      successful: true,
+    };
   });
